Fix: TC 17.01.08, is_valid_email method moved to base
